% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_preprocess.R
\name{alg_preprocess}
\alias{alg_preprocess}
\title{Preprocessing for multiple pairwise comparisons}
\usage{
alg_preprocess(
  x,
  alpha,
  measure,
  bootstrapping,
  nsample,
  severityRequirement,
  budget
)
}
\arguments{
\item{x}{a \code{m}-dimensional dataframe of the achieved objective values obtained for \code{n} runs by various algorithms}

\item{alpha}{is the significance level (\code{0< alpha <1}, default is 0.05)}

\item{measure}{indicates which measure to be compared among algorithms in the hypothesis testing, can take "mean", "median".}

\item{bootstrapping}{indicates if bootstrapping needs to be performed with hypothesis testing, default "FALSE",
accepts "TRUE" or "FALSE", choose "FALSE" for normal data-set and "TRUE" in-case of non-normal data-set}

\item{nsample}{the number of bootstrap samples, if needed}
}
\value{
returns a list of performance metrics
\itemize{
\item Main Algorithm- The algorithm which is tested to be outperforming
\item Other Compared Algorithm- The algorithm which is compared
\item P-Value- The P-value of the test
\item Adjusted P-Value- The Adjusted P-value of the test according to the methods of Benjamini, Hochberg, and Yekutieli to control the false discovery rate
}
}
\description{
The alg_preprocess is to obtain the p-value corrections for multiple pairwise comparisions
}
\author{
Sowmya Chandrasekaran
}
