% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_ranking.R
\name{alg_ranking}
\alias{alg_ranking}
\title{Ranking of the optimization algorithms based on mc}
\usage{
alg_ranking(df)
}
\arguments{
\item{df}{result of the alg_mcompare function, which is a data frame}
}
\value{
returns the algorithm ranked based on the score obtained
\itemize{
\item Algorithm- The algorithm which is tested to be outperforming
\item Score- Score obtained by each algorithm
\item Goal_Difference- Goal_Difference obtained by each algorithm
\item Ranking - Rank obtained by each algorithm
}
}
\description{
Ranking of the optimization algorithms based on mc
}
